compositor: introduce sub-surfaces
This commit depends on the Wayland commit "protocol: add sub-surfaces".

Implement the basic protocol for sub-surfaces:
- expose wl_subcompositor global interface
- error checking on protocol calls
- associate a parent wl_surface to a sub-surface
- introduce the sub-surface role, which is exclusive
- a rudimentary implementation of the wl_subsurface interface
- proper surface transformation inheritance from parent to sub-surfaces

Struct weston_subsurface is dynamically allocated. For sub-surfaces, it
is completely populated, and assigned to the weston_surface::private
pointer. Sub-surfaces are identified by the weston_surface::configure
function pointer.

For parent surfaces, weston_subsurface acts only as a link for stacking
order purposes. The wl_resource is unused, parent_destroy_listener is
not registered, and the transform is not linked.

Sub-surfaces are not added directly into layers for display or input.
Instead, a following patch will hook them up via the sub-surface list
present in parent weston_surfaces. This way sub-surfaces are inherently
linked to the parent surface, and cannot be displayed unless the parent
is mapped, too. This also eases restacking, as only the parent will be
in a layer list.

Features still lacking are:
- commit behaviour
	For wl_surface.commit, a parent and its sub-surfaces function
	still as independent surfaces.
- double buffered state
	Some state, like sub-surface position, is supposed to be applied
	on the next commit. For now, the state change is applied
	immediately.

Signed-off-by: Pekka Paalanen <ppaalanen@gmail.com>