compositor: add sub-surfaces to repaint list
Modify surface list rebuilding in weston_output_repaint() to process
sub-surface lists, if they are non-empty. The sub-surface list always
contains the parent, too, if not empty.

The collection of frame_callback_list is moved to a later loop, to
streamline the surface list rebuild functions. This should not change
anything.

Signed-off-by: Pekka Paalanen <ppaalanen@gmail.com>