shell: keyboard focus with sub-surfaces
The shell needs to redirect some actions to the parent surface, when
they originally target a sub-surface. This patch implements the
following:
- When the activation binding (left button click) targets a sub-surface,
  activate the parent surface instead. This also means, that a keyboard
  focus should never be on a sub-surface, but pointer (and touch) foci
  can.
- Move, resize, and rotate bindings always target the parent surface.
- Opacity (full-surface alpha) binding targets the parent surface. This
  is broken, because it should change the opacity of the whole compound
  window, which is difficult to implement in the renderer.

workspace_manager_move_surface(): is it expected and legal for this to
be called for sub-surfaces?

take_surface_to_workspace_by_seat(): I guess this really doesn't need
the weston_surface_get_parent call, since keyboard focus is supposed to
never be on a sub-surface.

move_surface_to_workspace(): like above, should be ok, as keyboard focus
is not a sub-surface.